python-app/requirements.txt
pkg-resources==0.0.0
PyMuPDF==1.16.10
python-dateutil==2.8.1
six==1.14.0